Guillaume : avancée ligneCommande dao et service

+++ Tp_Ecom/src/fr/adaming/model/LigneCommande.java
@@ -21,7 +21,7 @@ public class LigneCommande implements Serializable {
 	
 	private int quantite;
 	
-	private int prix;
+	private double prix;
 
 	
 	// Transformation de l'association UML en Java
@@ -69,11 +69,11 @@ public class LigneCommande implements Serializable {
 		this.quantite = quantite;
 	}
 
-	public int getPrix() {
+	public double getPrix() {
 		return prix;
 	}
 
-	public void setPrix(int prix) {
+	public void setPrix(double prix) {
 		this.prix = prix;
 	}
